% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{DIST_METHODS}
\alias{DIST_METHODS}
\title{Supported beta diversity distance methods}
\usage{
DIST_METHODS
}
\description{
Restricted to methods actually implemented in \code{.compute_dist()}
(data.R): most are passed straight to \code{vegan::vegdist()}, while
\code{"aitchison"} is computed via a CLR transform + Euclidean distance
and \code{"unifrac_weighted"} / \code{"unifrac_unweighted"} are computed
via \code{phyloseq::UniFrac()} using an uploaded phylogenetic tree.
}
