% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miome-package.R
\docType{package}
\name{miome-package}
\alias{miome}
\alias{miome-package}
\title{Miome: Microbiome Explorer}
\description{
Interactive Shiny app for PCoA, alpha diversity (LMM), taxonomy bar plots, beta diversity statistics, and MaAsLin3 differential abundance analysis of microbiome data.
}
\author{
\strong{Maintainer}: Your Name \email{your@email.com}

Authors:
\itemize{
  \item Your Name \email{your@email.com}
}

}
\keyword{internal}
