% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Launch the PCoA / Microbiome Explorer Shiny App}
\usage{
run_app()
}
\value{
Launches a Shiny app in your browser.
}
\description{
Launch the PCoA / Microbiome Explorer Shiny App
}
