% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{TAXA_DEFAULT_TOP_N}
\alias{TAXA_DEFAULT_TOP_N}
\title{Default number of top taxa shown individually in the taxonomy bar plot
before the remainder are collapsed into \code{"Others"}.}
\usage{
TAXA_DEFAULT_TOP_N
}
\description{
Default number of top taxa shown individually in the taxonomy bar plot
before the remainder are collapsed into \code{"Others"}.
}
